Add files via upload
As this theme was primarly quickly  done as is, simply as part of a tutorial and such theme .xml file was provided as part of the tutorial simply as an option if anyone had problems following such and needed to see such in a theme file. I opted to create a repo with the origanal txt file of the tutorial  and then simple change the theme dir name and theme name from dev as it was in the included tutorial to mybbcomlike for this repo theme .xml  just in case anyone needed it still as is for now it will be here in the repo and or if I ever decide to expand upon the tutorial or .xml file further such will be done via this repo of such.