1.8.33 usage
Upon a quick revisit to MyBB on latest version being now 1.8.33. A very simple edit to the included 1.8.31 xml theme file now for 1.8.33 usage of file that was originally included as an optional option to tutorial steps if folks had issues following such. Now moved the previous 1.8.31 theme file to archive folder in 1.8.31 directory in case such is needed for whatever reason